packages/python-sdk/pyproject.toml
[project]
name = "ambyte-sdk"
version = "0.1.0"
description = "The Python SDK for rights-aware data pipelines"
requires-python = ">=3.11"
dependencies = [
    "ambyte-schemas", # Depends on the local workspace package
    "opentelemetry-api>=1.38.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0", 
    "cachetools>=6.2.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
ambyte-schemas = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["ambyte"]